add support for several new decoding methods
1. topk, topp sampling and diverse beam search.
2. decoder support serving without encoder.

+++ example/transformer_example.cu.cc
@@ -65,8 +65,9 @@ int main(int argc, char *argv[]) {
   // instantiate decoder
   std::shared_ptr<byseqlib::cuda::Decoder<optype>> decoder_ =
       std::make_shared<byseqlib::cuda::Decoder<optype>>(
-          max_batch_size, reinterpret_cast<int *>(
-                              thrust::raw_pointer_cast(d_padding_mask_.data())),
+          max_batch_size,
+          reinterpret_cast<int *>(
+              thrust::raw_pointer_cast(d_padding_mask_.data())),
           reinterpret_cast<optraits::DataType *>(
               thrust::raw_pointer_cast(d_encoder_output_.data())),
           reinterpret_cast<int *>(thrust::raw_pointer_cast(d_output_.data())),
@@ -77,8 +78,8 @@ int main(int argc, char *argv[]) {
     return 1;
   }
   // init gpu memory buffer
-  int buf_bytesize = std::max(encoder_->compute_buffer_bytesize(),
-                              decoder_->compute_buffer_bytesize());
+  long buf_bytesize = std::max(encoder_->compute_buffer_bytesize(),
+                               decoder_->compute_buffer_bytesize());
   thrust::device_vector<int> d_buf_ =
       std::vector<int>(buf_bytesize / sizeof(int), 0);
   // encoder and decoder use the same buffer to save gpu memory useage
